+++ algorithms/src/test/java/com/design/analysis/quick/prep/array/IArrayTest.java
@@ -1,0 +1,14 @@
+package com.design.analysis.quick.prep.array;
+
+import org.junit.Before;
+
+public class IArrayTest {
+
+    IArray ir;
+
+    @Before
+    public void init()
+    {
+        ir = new ArrayImpl();
+    }
+}
