--- interview-corner/src/main/java/interview/preparation/self/basic/prep/IIntermediate.java
@@ -1,7 +1,0 @@
-package interview.preparation.self.basic.prep;
-//question for product company
-//https://www.geeksforgeeks.org/recently-asked-interview-questions-in-product-based-companies/
-public interface IIntermediate {
-
-	
-}
