+++ interview-corner/src/main/java/interview/preparation/self/basic/prep/prepdoc/hibernate.docx
